Move `copy_blockwise` and `nan_to_zero` to `_util.py`
